read me
In this project, I have developed a Style Advisor Recommender system using deep learning and image feature extraction techniques. The aim of this project is to recommend fashion items to users based on the visual similarity between uploaded images and a pre-trained ResNet50 model. Users can upload an image of a fashion item, and the recommender system will find and display the top 5 visually similar fashion items from a pre-defined dataset.